refactor[api-server]: lottery.parts 패키지 위치 변경
