+++ src/main/java/com/watermelon/server/event/lottery/parts/controller/PartsController.java
@@ -1,11 +1,11 @@
-package com.watermelon.server.lottery.parts.controller;
+package com.watermelon.server.event.lottery.parts.controller;
 
 import com.watermelon.server.event.lottery.auth.annotations.Uid;
-import com.watermelon.server.lottery.parts.dto.response.ResponseMyPartsListDto;
-import com.watermelon.server.lottery.parts.dto.response.ResponsePartsDrawDto;
-import com.watermelon.server.lottery.parts.dto.response.ResponseRemainChanceDto;
-import com.watermelon.server.lottery.parts.exception.PartsDrawLimitExceededException;
-import com.watermelon.server.lottery.parts.service.PartsService;
+import com.watermelon.server.event.lottery.parts.dto.response.ResponseMyPartsListDto;
+import com.watermelon.server.event.lottery.parts.dto.response.ResponsePartsDrawDto;
+import com.watermelon.server.event.lottery.parts.dto.response.ResponseRemainChanceDto;
+import com.watermelon.server.event.lottery.parts.exception.PartsDrawLimitExceededException;
+import com.watermelon.server.event.lottery.parts.service.PartsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
