+++ src/main/java/com/watermelon/server/event/lottery/parts/domain/PartsCategory.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.parts.domain;
+package com.watermelon.server.event.lottery.parts.domain;
 
 public enum PartsCategory {
     COLOR, REAR, DRIVE_MODE, WHEEL
