+++ src/main/java/com/watermelon/server/event/lottery/parts/exception/PartsDrawLimitExceededException.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.parts.exception;
+package com.watermelon.server.event.lottery.parts.exception;
 
 public class PartsDrawLimitExceededException extends RuntimeException {
 }
