+++ src/main/java/com/watermelon/server/event/lottery/parts/repository/PartsRepository.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.lottery.parts.repository;
+package com.watermelon.server.event.lottery.parts.repository;
 
-import com.watermelon.server.lottery.parts.domain.Parts;
+import com.watermelon.server.event.lottery.parts.domain.Parts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PartsRepository extends JpaRepository<Parts, Long> {
