+++ src/main/java/com/watermelon/server/event/lottery/parts/repository/PartsRewardRepository.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.lottery.parts.repository;
+package com.watermelon.server.event.lottery.parts.repository;
 
-import com.watermelon.server.lottery.parts.domain.PartsReward;
+import com.watermelon.server.event.lottery.parts.domain.PartsReward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PartsRewardRepository extends JpaRepository<PartsReward, Long> {
