+++ src/main/java/com/watermelon/server/event/lottery/parts/service/LotteryApplierPartsServiceImpl.java
@@ -1,9 +1,9 @@
-package com.watermelon.server.lottery.parts.service;
+package com.watermelon.server.event.lottery.parts.service;
 
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
-import com.watermelon.server.lottery.parts.domain.LotteryApplierParts;
-import com.watermelon.server.lottery.parts.domain.Parts;
-import com.watermelon.server.lottery.parts.repository.LotteryApplierPartsRepository;
+import com.watermelon.server.event.lottery.parts.domain.LotteryApplierParts;
+import com.watermelon.server.event.lottery.parts.domain.Parts;
+import com.watermelon.server.event.lottery.parts.repository.LotteryApplierPartsRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
