+++ src/main/java/com/watermelon/server/lottery/service/LotteryServiceImpl.java
@@ -10,8 +10,8 @@ import com.watermelon.server.event.lottery.dto.response.ResponseLotteryRankDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerInfoDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseRewardInfoDto;
-import com.watermelon.server.lottery.parts.domain.PartsReward;
-import com.watermelon.server.lottery.parts.repository.PartsRewardRepository;
+import com.watermelon.server.event.lottery.parts.domain.PartsReward;
+import com.watermelon.server.event.lottery.parts.repository.PartsRewardRepository;
 import com.watermelon.server.lottery.repository.LotteryApplierRepository;
 import com.watermelon.server.lottery.repository.LotteryRewardRepository;
 import jakarta.transaction.Transactional;
